Added the email alert when service_account keys are deleted (#213)
 * Improve precision of datetime comparisons for key expiration
 * Propagate notification errors in key rotate
 * Fix bug where expiring_keys was not checked before notification

+++ data_steward/admin/key_rotation.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import datetime, timedelta
 from oauth2client.client import GoogleCredentials
 from googleapiclient.errors import HttpError
 import googleapiclient.discovery as discovery
@@ -7,6 +7,8 @@ import logging
 LOGGER = logging.getLogger(__name__)
 
 KEY_EXPIRE_DAYS = 180
+KEY_EXPIRE_ALERT_DAYS = 7
+GCP_DTM_FMT = '%Y-%m-%dT%H:%M:%SZ'
 
 
 def get_iam_service():
@@ -41,26 +43,38 @@ def list_keys_for_service_account(service_account_email):
     return service_keys_per_account['keys']
 
 
-def is_key_expired(_key):
+def is_key_expired_after_period(key, days=KEY_EXPIRE_ALERT_DAYS):
+    """
+    Determine if a key will be expired after a specified number of days
+
+    :param key: service account key object
+    :param days: number of days
+    :return: True if the key is expiring soon based, False otherwise
+    """
+    now_dtm = datetime.now()
+    create_dtm = datetime.strptime(key['validAfterTime'], GCP_DTM_FMT)
+    expire_dtm = create_dtm + timedelta(days=KEY_EXPIRE_DAYS)
+    future_dtm = now_dtm + timedelta(days=days)
+    return future_dtm > expire_dtm
+
+
+def is_key_expired(key):
     """
     Determine if a key exceeds expiration period
 
-    :param _key: service account key object
+    :param key: service account key object
     :return: True if the key exceeds the expiration period, False otherwise
     """
-    today_date = datetime.today().date()
-    created_date = datetime.strptime(_key['validAfterTime'], '%Y-%m-%dT%H:%M:%SZ').date()
-    delta = today_date - created_date
-    return delta.days > KEY_EXPIRE_DAYS
+    return is_key_expired_after_period(key, 0)
 
 
-def delete_key(_key):
+def delete_key(key):
     """
     Delete a service account key
 
-    :param _key: service account key object
+    :param key: service account key object
     """
-    full_key_name = _key['name']
+    full_key_name = key['name']
     try:
         get_iam_service().projects().serviceAccounts().keys().delete(name=full_key_name).execute()
         LOGGER.info('{full_key_name} is deleted'.format(full_key_name=full_key_name))
@@ -77,12 +91,36 @@ def delete_expired_keys(project_id):
     Delete all expired service account keys associated with a project
 
     :param project_id: identifies the project
-    :return:
+    :return: a list of dicts that contain the info about the deleted keys
     """
-    for _service_account in list_service_accounts(project_id):
-        for key in list_keys_for_service_account(_service_account['email']):
+    deleted_keys = []
+    for service_account in list_service_accounts(project_id):
+        for key in list_keys_for_service_account(service_account['email']):
             if is_key_expired(key):
                 delete_key(key)
+                deleted_keys.append({'service_account_email': service_account['email'],
+                                     'key_name': key['name'],
+                                     'created_at': key['validAfterTime']})
+
+    return deleted_keys
+
+
+def get_expiring_keys(project_id):
+    """
+    List all expiring service account keys associated with a project
+
+    :param project_id: identifies the project
+    :return: a list of dicts that contain the info about the expiring keys
+    """
+    expiring_keys = []
+    for service_account in list_service_accounts(project_id):
+        for key in list_keys_for_service_account(service_account['email']):
+            if is_key_expired_after_period(key):
+                expiring_keys.append({'service_account_email': service_account['email'],
+                                      'key_name': key['name'],
+                                      'created_at': key['validAfterTime']})
+
+    return expiring_keys
 
 
 if __name__ == '__main__':
@@ -91,6 +129,18 @@ if __name__ == '__main__':
     PARSER = argparse.ArgumentParser(
         description='Delete all expired service account keys associated with a project',
         formatter_class=argparse.RawDescriptionHelpFormatter)
-    PARSER.add_argument('project_id', help='Identifies the project')
+    PARSER.add_argument('-p', '--project_id',
+                        help='Identifies the project',
+                        dest='project_id',
+                        required=True)
+    PARSER.add_argument('-d', '--delete',
+                        help='A flag to indicate whether or not to delete the keys',
+                        action='store_true')
     ARGS = PARSER.parse_args()
-    delete_expired_keys(ARGS.project_id)
+
+    if ARGS.delete:
+        DELETED_KEYS = delete_expired_keys(ARGS.project_id)
+    else:
+        EXPIRING_KEYS = get_expiring_keys(ARGS.project_id)
+        for KEY in EXPIRING_KEYS:
+            print(KEY)
